pca bug fix

+++ flaski/apps/main/pca.py
@@ -13,7 +13,7 @@ def make_figure(df,pa):
     if float( pa["percvar"].replace(",",".") ) < 100 :
         df_pca["__std__"]=df_pca.std(axis=1)
         df_pca=df_pca.sort_values( by=["__std__"],ascending=False )
-        nrows=round(len(df_pca)*float( pa["percvar"].replace(",",".") )/100)
+        nrows=int(round(len(df_pca)*float( pa["percvar"].replace(",",".") )/100))
         df_pca=df_pca[:nrows] 
         df_pca=df_pca.drop(["__std__"],axis=1)
